Moving system notification related files to SystemNotifications folder, removing tabs
