+++ src/SystemNotifications/SystemNotificationImageUnix.cpp
